requirements.txt
langchain-community
langchain-google-genai
langchain-tavily
langgraph

faiss-cpu

PyPDF2
langchain-text-splitters

pymongo
fastapi
uvicorn
gradio

requests
python-dotenv
tavily-python
